+++ vite.config.ts
@@ -1,12 +1,14 @@
-import { defineConfig } from "vite";
+import { defineConfig } from "vitest/config";
 import react from "@vitejs/plugin-react";
 
 export default defineConfig({
   plugins: [react()],
+  test: {
+    include: ["src/**/*.test.ts"]
+  },
   server: {
     strictPort: true,
     port: 5173
   },
   clearScreen: false
 });
-